[DEV-9734b] Remove color scale function
