+++ packages/chart/src/helpers/getColorScale.ts
@@ -2,54 +2,29 @@ import { colorPalettesChart as colorPalettes, twoColorPalette } from '@cdc/core/
 import { scaleOrdinal } from '@visx/scale'
 import { ChartConfig } from '../types/ChartConfig'
 
-const getColorScale = (config: ChartConfig, data: Object[]): ((value: string) => string) => {
-  const { visualizationSubType, visualizationType, series, legend } = config
+const getColorScale = (config: ChartConfig): ((value: string) => string) => {
+  const configPalette = ['Paired Bar', 'Deviation Bar'].includes(config.visualizationType)
+    ? config.twoColor.palette
+    : config.palette
+  const allPalettes: Record<string, string[]> = { ...colorPalettes, ...twoColorPalette }
+  let palette = config.customColors || allPalettes[configPalette]
+  let numberOfKeys = config.runtime.seriesKeys.length
+  let newColorScale
 
-  const generatePalette = colorsCount => {
-    if (!series?.length) return []
-    const isSpecialType = ['Paired Bar', 'Deviation Bar'].includes(visualizationType)
-    const chosenPalette = isSpecialType ? config.twoColor.palette : config.palette
-    const allPalettes = { ...colorPalettes, ...twoColorPalette }
-    let palette = config.customColors || allPalettes[chosenPalette]
-    while (colorsCount > palette.length) palette = palette.concat(palette)
-    return palette.slice(0, colorsCount)
+  while (numberOfKeys > palette.length) {
+    palette = palette.concat(palette)
   }
 
-  let colorScale = scaleOrdinal<string, string>({
-    domain: config?.runtime?.seriesLabelsAll,
-    range: generatePalette(series.length),
-    unknown: null
-  })
+  palette = palette.slice(0, numberOfKeys)
 
-  if (visualizationType === 'Deviation Bar') {
-    const { targetLabel } = config.xAxis
-    colorScale = scaleOrdinal<string, string>({
-      domain: [`Below ${targetLabel}`, `Above ${targetLabel}`],
-      range: generatePalette(2),
+  newColorScale = () =>
+    scaleOrdinal({
+      domain: config.runtime.seriesLabelsAll,
+      range: palette,
       unknown: null
     })
-  }
-  if (visualizationType === 'Bar' && visualizationSubType === 'regular' && series?.length === 1 && legend?.colorCode) {
-    const set = new Set(data?.map(d => d[legend.colorCode]))
-    const domain = Array.from(set)
-    colorScale = scaleOrdinal<string, string>({
-      domain: domain,
-      range: generatePalette([...domain].length),
-      unknown: null
-    })
-  }
-  if (config.series.some(s => s.name)) {
-    const set = new Set(series.map(d => d.name || d.dataKey))
-    const domain = Array.from(set)
 
-    colorScale = colorScale = scaleOrdinal<string, string>({
-      domain: [...domain],
-      range: generatePalette(series.length),
-      unknown: null
-    })
-  }
-
-  return colorScale
+  return newColorScale
 }
 
 export default getColorScale
